refactor autorepeat patch to diff package
