+++ src/main/kotlin/app/revanced/patches/youtube/misc/autorepeat/fingerprints/AutoRepeatParentFingerprint.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.youtube.layout.autorepeat.fingerprints
+package app.revanced.patches.youtube.misc.autorepeat.fingerprints
 
 import app.revanced.patcher.annotation.Name
 import app.revanced.patcher.annotation.Version
@@ -6,7 +6,7 @@ import app.revanced.patcher.extensions.or
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
 import app.revanced.patcher.fingerprint.method.annotation.FuzzyPatternScanMethod
 import app.revanced.patcher.fingerprint.method.annotation.MatchingMethod
-import app.revanced.patches.youtube.layout.autorepeat.annotations.AutoRepeatCompatibility
+import app.revanced.patches.youtube.misc.autorepeat.annotations.AutoRepeatCompatibility
 import org.jf.dexlib2.AccessFlags
 
 @Name("auto-repeat-parent-fingerprint")
